feature. move kubernetes pkg
