+++ api/swagger/swagger.json
@@ -12220,7 +12220,8 @@
                     "$ref": "#/definitions/github_com_openinfradev_tks-api_pkg_domain.Match"
                 },
                 "matchYaml": {
-                    "type": "string"
+                    "type": "string",
+                    "example": "namespaces:\r\n- testns1"
                 },
                 "parameters": {
                     "type": "string",
